增加订单页面（如果本地没有订单数据可能无法正常展示）,并初步实现支付和显示二维码的功能，但由于没有同步到后端，所以数据并不会被保存

+++ storage/userinfo.h
@@ -9,7 +9,7 @@ class UserInfo : public QObject
     Q_OBJECT
     Q_PROPERTY(QString userName READ userName WRITE setUserName NOTIFY userNameChanged)
     Q_PROPERTY(QString userPersonalInfo READ userPersonalInfo WRITE setUserPersonalInfo NOTIFY userPersonalInfoChanged)
-    Q_PROPERTY(int myMoney READ myMoney WRITE setMyMoney NOTIFY myMoneyChanged)
+    Q_PROPERTY(double myMoney READ myMoney WRITE setMyMoney NOTIFY myMoneyChanged)
     Q_PROPERTY(QString myToken READ myToken WRITE setMyToken NOTIFY myTokenChanged)
     Q_PROPERTY(QString userEmail READ userEmail WRITE setUserEmail NOTIFY userEmailChanged)
     Q_PROPERTY(QString myAvatar READ myAvatar WRITE setMyAvatar NOTIFY myAvatarChanged)
@@ -87,7 +87,7 @@ signals:
 private:
     QString m_userName;
     QString m_userPersonalInfo;
-    int m_myMoney;
+    double m_myMoney;
     QString m_myTelephone;
     QString m_myToken;
     int m_userId;
